+++ src/vroom_date.h
@@ -11,7 +11,8 @@ double parse_date(
     const char* begin,
     const char* end,
     DateTimeParser& parser,
-    const std::string& format);
+    const std::string& format,
+    LocaleInfo* locale);
 
 cpp11::doubles read_date(vroom_vec_info* info);
 
@@ -68,7 +69,7 @@ public:
         i,
         info->info->column,
         [&](const char* begin, const char* end) -> double {
-          return parse_date(begin, end, *info->parser, info->info->format);
+          return parse_date(begin, end, *info->parser, info->info->format, info->info->locale.get());
         },
         info->info->errors,
         err_msg.c_str(),